+++ authentication/README.md
@@ -1,0 +1,3 @@
+# Authentication and Invites
+
+For documentation visit https://top-caldron-366821.ew.r.appspot.com/docs
